@@ -9,11 +9,11 @@
   },
   "devDependencies": {
     "@types/node": "^20.12.7",
-    "@typescript-eslint/eslint-plugin": "^5.59.2",
-    "@typescript-eslint/parser": "^5.59.2",
-    "eslint": "^8.39.0",
-    "tsx": "^4.6.2",
-    "typescript": "^5.3.3"
+    "@typescript-eslint/eslint-plugin": "^7.8.0",
+    "@typescript-eslint/parser": "^7.8.0",
+    "eslint": "^8.57.0",
+    "tsx": "^4.9.1",
+    "typescript": "^5.4.5"
   },
   "dependencies": {
     "@js-joda/core": "^5.5.3",
