@@ -8,5 +8,5 @@ export function check(client1: Db, client2: Db) {
   client1.queryOne<{ post_id: number | null }>(sql`SELECT post_id FROM comments`);
 
   // client 2 points to acme/migrations2/
-  client2.queryOne<{ name: string | null }>(sql`SELECT name  FROM chat_rooms`);
+  client2.queryOne<{ name: string | null }>(sql`SELECT name FROM chat_rooms`);
 }
