@@ -1,18 +1,14 @@
-import { execSync } from "child_process";
-import { config } from "dotenv";
-import path from "path";
-
-config({ path: path.join(__dirname, "..", ".env") });
+import { $ } from "zx";
 
 const DATABASE_NAME = "safeql_prisma_demo";
 
 async function main() {
-  execSync(`psql -c "DROP DATABASE IF EXISTS ${DATABASE_NAME} WITH (FORCE);"`);
-  execSync(`psql -U postgres -c "CREATE DATABASE ${DATABASE_NAME};"`);
-  execSync(`pnpm prisma generate`, { stdio: "inherit" });
-  execSync(`pnpm prisma migrate deploy`, { stdio: "inherit" });
+  await $`psql -c "DROP DATABASE IF EXISTS ${DATABASE_NAME} WITH (FORCE);"`;
+  await $`psql -U postgres -c "CREATE DATABASE ${DATABASE_NAME};"`;
+  await $`pnpm prisma generate`;
+  await $`pnpm prisma migrate deploy`;
 
-  const { PrismaClient } = await import("@prisma/client");
+  const { PrismaClient } = await import("@/prisma-client");
   const prisma = new PrismaClient();
 
   await prisma.user.createMany({
