@@ -1,3 +1,4 @@
+import { enhanceAppWithTabs } from "vitepress-plugin-tabs/client";
 import Theme from "vitepress/theme";
 import Layout from "./Layout.vue";
 import "./custom.css";
@@ -5,7 +6,9 @@ import "./custom.css";
 export default {
   extends: Theme,
   Layout: Layout,
-  async enhanceApp() {
+  async enhanceApp({ app }) {
+    enhanceAppWithTabs(app);
+
     import("@vercel/analytics").then(({ default: analytics }) => {
       analytics.inject();
     });
