@@ -1,6 +1,5 @@
-export { json } from "fp-ts";
-export * as E from "fp-ts/Either";
-export { flow, identity, pipe } from "fp-ts/function";
-export * as O from "fp-ts/Option";
-export * as TE from "fp-ts/TaskEither";
-export * as J from "fp-ts/Json";
+export * as E from "fp-ts/lib/Either.js";
+export { flow, identity, pipe } from "fp-ts/lib/function.js";
+export * as O from "fp-ts/lib/Option.js";
+export * as TE from "fp-ts/lib/TaskEither.js";
+export * as J from "fp-ts/lib/Json.js";
