@@ -1,21 +1,13 @@
 import chokidar from "chokidar";
 import path from "path";
 import { z } from "zod";
-import { RuleOptionConnection, zConnectionMigration } from "../rules/check-sql.rule";
+import { RuleOptionConnection, zConnectionMigration } from "../rules/RuleOptions";
 import {
   DEFAULT_CONNECTION_URL,
   getDatabaseName,
   getMigrationDatabaseMetadata,
 } from "../rules/check-sql.utils";
 import { CloseConnectionParams } from "./connection-manager";
-
-export function createWatchManager() {
-  const watchers: Map<string, unknown> = new Map();
-
-  return {
-    watchMigrationsDir: (params: WatchMigrationsDirParams) => watchMigrationsDir(params, watchers),
-  };
-}
 
 interface WatchMigrationsDirParams {
   projectDir: string;
@@ -24,29 +16,44 @@ interface WatchMigrationsDirParams {
   closeConnectionFn: (params: CloseConnectionParams) => void;
 }
 
-export function watchMigrationsDir(
-  params: WatchMigrationsDirParams,
-  watchers: Map<string, unknown>
-) {
-  const migrationsPath = path.join(params.projectDir, params.connection.migrationsDir);
-
-  if (watchers.has(migrationsPath)) {
-    return;
-  }
-
-  const { databaseUrl } = getMigrationDatabaseMetadata({
-    connectionUrl: params.connection.connectionUrl ?? DEFAULT_CONNECTION_URL,
-    databaseName: getDatabaseName({
-      databaseName: params.connection.databaseName,
-      migrationsDir: params.connection.migrationsDir,
-      projectDir: params.projectDir,
-    }),
+export function createWatchManager() {
+  const migrationPaths = new Map<string, { onChange: () => void }>();
+  const watcher = chokidar.watch([], { ignoreInitial: true }).on("all", (_, filePath) => {
+    for (const [path, { onChange }] of migrationPaths.entries()) {
+      if (filePath.startsWith(path)) {
+        return onChange();
+      }
+    }
   });
 
-  const watcher = chokidar.watch(migrationsPath, { ignoreInitial: true }).on("all", () => {
-    params.dropCacheKeyFn(databaseUrl);
-    params.closeConnectionFn({ connection: params.connection, projectDir: params.projectDir });
-  });
+  const watchMigrationsDir = (params: WatchMigrationsDirParams) => {
+    const migrationPath = path.join(params.projectDir, params.connection.migrationsDir);
 
-  watchers.set(migrationsPath, watcher);
+    if (migrationPaths.has(migrationPath)) {
+      return;
+    }
+
+    migrationPaths.set(migrationPath, {
+      onChange: () => {
+        const { databaseUrl } = getMigrationDatabaseMetadata({
+          connectionUrl: params.connection.connectionUrl ?? DEFAULT_CONNECTION_URL,
+          databaseName: getDatabaseName({
+            databaseName: params.connection.databaseName,
+            migrationsDir: params.connection.migrationsDir,
+            projectDir: params.projectDir,
+          }),
+        });
+
+        params.dropCacheKeyFn(databaseUrl);
+        params.closeConnectionFn({ connection: params.connection, projectDir: params.projectDir });
+      },
+    });
+
+    console.log("migrationPath", migrationPath);
+    watcher.add(migrationPath);
+  };
+
+  return {
+    watchMigrationsDir,
+  };
 }
