@@ -1,8 +1,8 @@
-export {
+export type {
   GenerateError,
   GenerateParams,
   GenerateResult,
   ResolvedTargetEntry,
   ResolvedTarget,
-  createGenerator,
 } from "./src/generate";
+export { createGenerator } from "./src/generate";
